Add markets filter on
